Rename files with shared examples
Files with Shared examples for Concerns Votable and Voted should not end
with _spec, this causes warning 'Examples Previously defined'
