+++ spec/controllers/concerns/voted.rb
