+++ spec/models/concerns/votable.rb
